projects/02_trivia_api/starter/backend/requirements.txt
aniso8601==6.0.0
Click==7.0
Flask==1.0.3
flask_cors==3.0.7
Flask-RESTful==0.3.7
flask_sqlalchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2-binary
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.4
Werkzeug
babel
python-dateutil==2.6.0
flask-wtf
python-dotenv
